Merge development to master

+++ app/user_service/models.py
@@ -47,7 +47,7 @@ class UserDataAccess:
         pass
 
     def get_users(self):
-        rows = db.engine.execute('SELECT Username, Pass, FirstName, LastName, Email, Status, Active FROM Member;')
+        rows = db.engine.execute('SELECT * FROM Member;')
         quote_objects = list()
         for row in rows:
             quote_obj = User(row['username'], row['pass'], row['firstname'], row['lastname'], row['email'],
@@ -55,12 +55,27 @@ class UserDataAccess:
             quote_objects.append(quote_obj)
         return quote_objects
 
+    def get_admins(self):
+        """ Returns a list of users that are admins """
+        try:
+            rows = db.engine.execute("SELECT * FROM Member WHERE Status = 'admin';")
+            admins = list()
+            for row in rows:
+                admin = User(row['username'], row['pass'], row['firstname'], row['lastname'], row['email'],
+                                 row['status'], row['active'])
+                admins.append(admin)
+            return admins
+        except Exception as e:
+            app.logger.error("[ERROR] Unable to fetch admin list.")
+            app.logger.exception(e)
+            raise e
+
     def add_user(self, user_obj):
         try:
-            query = 'INSERT INTO Member(Username,Pass,FirstName,LastName,Email,Status,Active) VALUES({},{},{},{},{},{},{})'.format(
+            query = "INSERT INTO Member VALUES({},{},{},{},{},{},{}) ON CONFLICT (Username) WHERE Username={} DO NOTHING;".format(
                 *_cv(
                     user_obj.username, user_obj.password, user_obj.firstname, user_obj.lastname, user_obj.email,
-                    user_obj.status), user_obj.is_active)
+                    user_obj.status), user_obj.is_active, _cv(app.config['ADMIN_USERNAME']))
             db.engine.execute(query)
             return True
         except Exception as e:
@@ -81,14 +96,11 @@ class UserDataAccess:
         rows = db.engine.execute(
             'SELECT * FROM Member WHERE Username={};'.format(_cv(user_id)))
         row = rows.first()
-
-        if row is not None:
-            user = User(row['username'], row['pass'], row['firstname'], row['lastname'], row['email'],
-                        row['status'], row['active'])
-            return user
-        else:
-            raise Exception
-            return None
+        try:
+            return User(row['username'], row['pass'], row['firstname'], row['lastname'],
+                        row['email'], row['status'], row['active'])
+        except TypeError as e:
+            raise Exception("Failed to get user.") # This could/should be a custom exception
 
     def alter_user(self, user):
         try:
@@ -100,18 +112,38 @@ class UserDataAccess:
             db.engine.execute(query)
             return True
         except Exception as e:
+            app.logger.error("[ERROR] Unable to alter user.")
             app.logger.exception(e)
-            raise Exception
+            raise e
+
+    def set_admin(self, username, admin=True):
+        """ Sets the given users admin status """
+        # Can't change status of main admin
+        if username == app.config['ADMIN_USERNAME']:
+            raise "Cannot change status of main admin."
+        try:
+            user = self.get_user(username)
+            user.status = 'admin' if admin else 'user'
+            self.alter_user(user)
+        except Exception as e:
+            app.logger.error("[ERROR] Unable set admin on user.")
+            app.logger.exception(e)
+            raise e
 
     def delete_user(self, data_loader, username):
         """remove user and all of its datasets"""
+        # Don't allow main admin deletion
+        if username == app.config['ADMIN_USERNAME']:
+            return False
+
         # remove user deletes every row that depends on it because of cascade deletion
         try:
             # first drop all schemas owned by the user
             query = 'SELECT id FROM dataset WHERE owner = {}'.format(_cv(username))
             rows = db.engine.execute(query)
             for dataset_id in rows:
-                data_loader.delete_dataset(dataset_id[0])
+                schema_id = dataset_id[0].split('-')[1]
+                data_loader.delete_dataset(schema_id)
             db.engine.execute('DELETE FROM Member WHERE username = {}'.format(_cv(username)))
             return True
         except Exception as e:
